Modify NetworkReconciler to handle termination of networks and update go.mod file
Integrate a mechanism to manage and handle the termination process of networks in the NetworkReconciler. The added feature ensures that Virtual Network Interface (VNI) resources are not released prematurely, but only after the termination procedure is completed. This improvement aids in enhancing the strength and accuracy of the reconciliation process. The deletion flow procedure has been revamped and a queuing system for retrying finalizer removal has been implemented. Also, the go.mod file has been updated systemically.